Update 2-0-0
Temporary, testing branch to get a dev environment Postman collection working. Will eventually move these changes to a different, permanent branch once we have a working proof of concept; for now, just need a branch with a schema that's linkable for the Snowplow event requests.

Changed the self-referencing schema on line 2 to the URL for this schema, and removed the encoding option from the 'elastic_search_query' property.